chore: split up filter components
